% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum-functions.R
\name{fm_graph}
\alias{fm_graph}
\title{fm graph}
\usage{
fm_graph(x, merge_edges = TRUE)
}
\arguments{
\item{x}{A data frame. Needs the fields 'poster' and 'reply_to',
and also currently 'hit_count', 'posted_date' and 'msg_text'}
}
\description{
Makes a graph from raw forum output.
}
