% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum-functions.R
\name{plot3d_graph}
\alias{plot3d_graph}
\title{plot3d_graph}
\usage{
plot3d_graph(graph)
}
\arguments{
\item{graph}{A network object}
}
\description{
Takes a network object and plots a 3d hairball plot.
}
